+++ GenericModConfigMenu/Framework/OwnModConfig.cs
@@ -1,3 +1,5 @@
+using StardewModdingAPI.Utilities;
+
 namespace GenericModConfigMenu.Framework
 {
     /// <summary>The mod configuration for Generic Mod Config Menu itself.</summary>
@@ -8,5 +10,8 @@ namespace GenericModConfigMenu.Framework
         *********/
         /// <summary>The number of field rows to offset when scrolling a config menu.</summary>
         public int ScrollSpeed { get; set; } = 120;
+
+        /// <summary>A keybind which opens the menu.</summary>
+        public KeybindList OpenMenuKey = new KeybindList(StardewModdingAPI.SButton.None);
     }
 }