+++ GenericModConfigMenu/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Added optional keybind to open the config UI.
+
 ## 1.6.0
 Released 29 October 2021 for SMAPI 3.12.5 or later. Updated by Pathoschild.